requirements.txt
dotenv
requests
pymongo
schedule
python-dotenv
kafka-python
cassandra-driver